% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mccann_spike.R
\name{add_mccann_spike}
\alias{add_mccann_spike}
\title{Add McCann spike to data}
\usage{
add_mccann_spike(x, spikes = c(49, 64, 74, 84))
}
\arguments{
\item{x}{totals}

\item{spikes}{place to add one}
}
\value{
new totals with no one off a boundary
}
\description{
Add McCann spike to data
}
\examples{
add_mccann_spike(1:100)
add_mccann_spike(NA)
}
