% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_marks.R
\name{clean_marks_df}
\alias{clean_marks_df}
\title{Takes a data frame of marks and regular expression and converts 
the columns whose names match regular expression}
\usage{
clean_marks_df(df, assess_RE)
}
\arguments{
\item{df}{data frame of marks}

\item{RE}{regular expression used to identify the columns}
}
\value{
data frame with exemptions converted to value.
}
\description{
Takes a data frame of marks and regular expression and converts 
the columns whose names match regular expression
}
\examples{
library(examMarking)
library(examMarking)
data("SMI_2018_marks")
SMI_2018_marks
clean_marks_df(SMI_2018_marks, "^A")
}
