% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_calculate_mean.R
\name{exam_calculate_mean}
\alias{exam_calculate_mean}
\title{takes data frame and regular expression and calculates the mean for the given 
columns. Will take into account exemmptions}
\usage{
exam_calculate_mean(df, RE)
}
\arguments{
\item{df}{data frame of marks}

\item{RE}{regular expression to choose columns to average}
}
\value{
vector of means
}
\description{
takes data frame and regular expression and calculates the mean for the given 
columns. Will take into account exemmptions
}
\examples{
raw_data \%>\% 
  exam_canvas_clean("^A") \%>\% 
  exam_calculate_prop("^A", 1) \%>\% 
  exam_calculate_mean("^A")
}
