% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_calculate_prop.R
\name{exam_calculate_prop}
\alias{exam_calculate_prop}
\title{Get proportions for dataframe.}
\usage{
exam_calculate_prop(df, assess_RE, total_row)
}
\arguments{
\item{df}{data frame of marks}

\item{assess_RE}{regular expression to choose columns}
}
\value{
data frame with proportions
}
\description{
Identifies columns by regular expression
}
\examples{
data("SMI_2018_marks")
SMI_2018_marks  <-
  SMI_2018_marks \%>\%
  exam_canvas_clean("^A")
SMI_2018_marks
SMI_2018_marks \%>\%
  exam_calculate_prop("^A", 1)
}
