% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_calculate_total.R
\name{exam_calculate_total}
\alias{exam_calculate_total}
\title{takes data frame and regular expression and calculates the totals for the given 
columns}
\usage{
exam_calculate_total(df, RE)
}
\arguments{
\item{df}{data frame of marks}

\item{RE}{regular expression to choose columns to total}
}
\value{
vector of totals
}
\description{
takes data frame and regular expression and calculates the totals for the given 
columns
}
