% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_canvas_clean.R
\name{exam_canvas_clean}
\alias{exam_canvas_clean}
\title{Takes a data frame of marks and regular expression and converts 
the columns whose names match regular expression}
\usage{
exam_canvas_clean(df, assess_RE)
}
\arguments{
\item{df}{data frame of marks}

\item{RE}{regular expression used to identify the columns}
}
\value{
data frame with exemptions converted to value.
}
\description{
Takes a data frame of marks and regular expression and converts 
the columns whose names match regular expression
}
\examples{
data("raw_data")
SMI_2018_marks
exam_canvas_clean(SMI_2018_marks, "^A")
}
