% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_clean_grade_book.R
\name{exam_clean_grade_book}
\alias{exam_clean_grade_book}
\title{exam_clean_grade_book}
\usage{
exam_clean_grade_book(df, cols)
}
\arguments{
\item{df}{a data frame from MyUni gradebook}

\item{cols}{tidyselect columns to clean}
}
\value{
cleaned dataframe
}
\description{
exam_clean_grade_book
}
