% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_canvas_clean.R
\name{exam_clean_marks}
\alias{exam_clean_marks}
\title{takes a vector of marks and converts missing to zero and EX to NA}
\usage{
exam_clean_marks(x)
}
\arguments{
\item{x}{marks}
}
\value{
vector of marks
}
\description{
takes a vector of marks and converts missing to zero and EX to NA
}
