% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_compare_norm.R
\name{exam_compare_norm}
\alias{exam_compare_norm}
\title{Compare raw and norm marks}
\usage{
exam_compare_norm(df, raw, norm, first_year = FALSE)
}
\arguments{
\item{df}{marks data frame}

\item{raw}{column of raw marks}

\item{norm}{column of norm marks}

\item{...}{grading parameters}
}
\value{
table comparing proportions
}
\description{
Compare raw and norm marks
}
\examples{
marks \%>\% 
mutate(norm = exam_norm_glonek(total, a = 1, b = 1.1)) \%>\% 
  exam_compare_norm(total, norm, first_year = "TRUE")
}
