% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_convert_prop.R
\name{exam_convert_prop}
\alias{exam_convert_prop}
\title{exam_convert_prop}
\usage{
exam_convert_prop(df, cols, row)
}
\arguments{
\item{df}{a dataframe with totals in one row}

\item{cols}{tidyselection of columns}

\item{row}{number of row with totals}
}
\value{
dataframe with proportions
}
\description{
given a dataframe, some columns and a total row, converts
to proportions
}
\examples{
pacman::p_load(tidyverse, targets)
df <- tibble(
  A1 = c(10, 1, 2),
  A2 = c(10, 2, 1)
)
df |> exam_convert_prop(A1, 1)
df |> exam_convert_prop(A1:A2, 1)
}
