% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam-csv-xlsx.r
\name{exam_csv_xlsx}
\alias{exam_csv_xlsx}
\title{Converts CSV to XLSX}
\usage{
exam_csv_xlsx(CSV)
}
\arguments{
\item{CSV}{filename of CSV}
}
\value{
filename of XLSX
}
\description{
Converts CSV to XLSX
}
\examples{
exam_csv_xlsx("inst/extdata/example-grade-roster-filled.csv")
}
