% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_fill_grade_roster.R
\name{exam_fill_grade_roster}
\alias{exam_fill_grade_roster}
\title{fill_grade_roster}
\usage{
exam_fill_grade_roster(df, IDs, total, infile, outfile, supp = 45:49,
  supp_code = "US10", RP = NULL, trace = TRUE)
}
\arguments{
\item{df}{data frame of marks}

\item{IDs}{column of IDs}

\item{total}{column with rounded totals}

\item{infile}{empty grade roster}

\item{outfile}{lace to save complete grade roster}
}
\value{
VOID
}
\description{
Still need to open, check and save
}
\examples{
pacman::p_load(tidyverse)
set.seed(2019)
df  <- tibble(
  ID = 1:7,
  total = c(sample(80:100, 4), 0, 49, 44)
)
df
df \%>\% exam_fill_grade_roster(ID, total,
                  infile = "inst/grade_example.csv",
                  outfile = "inst/example_output.csv")
read_lines("inst/example_output.csv")
df \%>\% exam_fill_grade_roster(ID, total,
                  infile = "inst/grade_example.csv",
                  outfile = "inst/example_output_2.csv",
                  supp = 40:49)
read_lines("inst/example_output_2.csv")
df \%>\% exam_fill_grade_roster(ID, total,
                  infile = "inst/grade_example.csv",
                  outfile = "inst/example_output_3.csv",
                  RP = c(2, 5))
read_lines("inst/example_output_3.csv")
}
\author{
Jono Tuke
}
