% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_get_email_list.R
\name{exam_get_email_list}
\alias{exam_get_email_list}
\title{Get list of student emails from IDs}
\usage{
exam_get_email_list(IDs, form = "a{IDs}@student.adelaide.edu.au")
}
\arguments{
\item{IDs}{vector of IDs of form a1078881 or 1078881}

\item{form}{glue string for email address}
}
\value{
single string with all emails seperated by a comma
}
\description{
Get list of student emails from IDs
}
\examples{
IDs  <- c('a1078881', 'a1078886')
exam_get_email_list(IDs)
}
