% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_get_enrolled.R
\name{exam_get_enrolled}
\alias{exam_get_enrolled}
\title{get enrolled}
\usage{
exam_get_enrolled(file)
}
\arguments{
\item{file}{grade roster}
}
\value{
tibble of enrolled
}
\description{
uses the grade roster to get enrolled students.
}
\examples{
exam_get_enrolled(file = "inst/grade_example.csv")
}
