% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_get_grades.R
\name{exam_get_grade}
\alias{exam_get_grade}
\title{convert integer marks to grades}
\usage{
exam_get_grade(marks, grade_tab = NULL, trace = TRUE, service = FALSE)
}
\arguments{
\item{marks}{vector of integer marks}

\item{grades}{table of grade boundaries, needs cols: grade, lower, upper}
}
\value{
grades
}
\description{
uses a table of upper and lower points to assign
grades
}
\examples{
tibble(
  marks = 0:100,
  grade = exam_get_grade(marks, trace = TRUE)
  ) \%>\%
  ggplot(aes(marks, grade, col = grade)) +
  geom_point() +
  scale_x_continuous(breaks = seq(0, 100, 5))
  exam_get_grade(0:100, service = TRUE)
  exam_get_grade(0:100, service = FALSE)
}
