% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_get_prop.R
\name{exam_get_prop}
\alias{exam_get_prop}
\title{exam_get_prop}
\usage{
exam_get_prop(x, total)
}
\arguments{
\item{x}{numeric vector}

\item{total}{index of total entry}
}
\value{
proportions
}
\description{
exam_get_prop
}
\examples{
x <- 1:10
exam_get_prop(x, 10)
}
