% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_grade.R
\name{exam_grade}
\alias{exam_grade}
\title{convert integer marks to grades}
\usage{
exam_grade(marks, grade_tab = NULL, trace = TRUE, first_year = FALSE)
}
\arguments{
\item{marks}{vector of integer marks}

\item{grades}{table of grade boundaries, needs cols: grade, lower, upper}
}
\value{
grades
}
\description{
uses a table of upper and lower points to assign
grades
}
\examples{
tibble(marks = 0:100, grade = exam_grade(marks, trace = TRUE)) \%>\% 
  ggplot(aes(marks, grade, col = grade)) + 
  geom_point() + 
  geom_line(size = 4) + 
  scale_x_continuous(breaks = seq(0, 100, 5))
  exam_grade(0:100, first_year = TRUE)
  exam_grade(0:100, first_year = FALSE)
}
