% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_grade_summary.R
\name{exam_grade_summary}
\alias{exam_grade_summary}
\title{Takes column of grades and gives summary}
\usage{
exam_grade_summary(df, grade_col)
}
\arguments{
\item{grades}{vector of grades}
}
\value{
table of counts and prop.
}
\description{
Takes column of grades and gives summary
}
\examples{
marks \%>\% exam_grade_summary(grade)
}
