% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_norm_glonek.R
\name{exam_norm_glonek}
\alias{exam_norm_glonek}
\title{Normalizes using Glonek function}
\usage{
exam_norm_glonek(x, a = 0, b = 1)
}
\arguments{
\item{x}{marks}

\item{a}{normalization parameter}

\item{b}{normalization parameter}
}
\value{
normalized marks
}
\description{
Takes marks and normalizes using Glonek function - linear
scale on logit scale
}
\examples{
df  <- tibble(
  raw = 0:100, 
  `a = 0.5, b = 1.1` = exam_norm_glonek(raw, 0.5, 1.1),
  `a = 0.5, b = 1.5` = exam_norm_glonek(raw, 0.5, 1.5),
  `a = 0.5, b = 0.9` = exam_norm_glonek(raw, 0.5, 0.9)
)
df \%>\% 
  pivot_longer(-raw) \%>\% 
  ggplot(aes(raw, value, col = name)) + 
  geom_point() + 
  geom_abline(intercept = 0, slope = 1)
}
\author{
Jono Tuke <simon.tuke@adelaide.edu.au>
}
