% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_norm_mccann.R
\name{exam_norm_mccann}
\alias{exam_norm_mccann}
\title{Add McCann spike to data}
\usage{
exam_norm_mccann(x, spikes = c(49, 64, 74, 84))
}
\arguments{
\item{x}{totals}

\item{spikes}{place to add one}
}
\value{
new totals with no one off a boundary
}
\description{
Add McCann spike to data
}
\examples{
exam_norm_mccann(1:100)
exam_norm_mccann(NA)
}
