% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_plot_assessment.R
\name{exam_plot_assessment}
\alias{exam_plot_assessment}
\title{Plots the assessment for each student}
\usage{
exam_plot_assessment(df, RE, name, grade)
}
\arguments{
\item{df}{data frame of course}

\item{RE}{regular expression of section totals}

\item{grades}{vector of grades}

\item{names}{names of students}
}
\value{
a ggplot
}
\description{
Only good for small coursese
}
\examples{
data("SMI_2018_marks_clean")
SMI_2018_marks_clean \%>\% 
exam_plot_assessment(RE = "^A$|^OQ$|^P$|^E$", name = ID, grade = grade)
}
