% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_plot_grades.R
\name{exam_plot_grades}
\alias{exam_plot_grades}
\title{Get grade plot}
\usage{
exam_plot_grades(df, grade_col)
}
\arguments{
\item{df}{data frame of marks}

\item{grade_col}{column with grades}
}
\value{
ggplot
}
\description{
Get grade plot
}
\examples{
marks \%>\% exam_plot_grades(grade)
}
