% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_plot_norm.R
\name{exam_plot_norm}
\alias{exam_plot_norm}
\title{Plot raw and norm}
\usage{
exam_plot_norm(df, raw, norm)
}
\arguments{
\item{df}{data frame of data}

\item{raw}{raw marks column}

\item{norm}{normalised marks column}
}
\value{
ggplot
}
\description{
Plot raw and norm
}
\examples{
marks \%>\% 
mutate(norm = round(exam_norm_glonek(total, a = 0.2, b = 1))) \%>\% 
  exam_plot_norm(total, norm) 
}
