% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_plot_profile.R
\name{exam_plot_profile}
\alias{exam_plot_profile}
\title{Plots the exam questions to look for seperation}
\usage{
exam_plot_profile(df, RE, grade_col)
}
\arguments{
\item{df}{data frame with results}

\item{RE}{regular expression of exam questions}

\item{grade}{column with grades}
}
\value{
a ggplot
}
\description{
Plots the exam questions to look for seperation
}
\examples{
marks \%>\% exam_plot_profile("^Q", grade)
}
