% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_xlsx_csv.r
\name{exam_xlsx_csv}
\alias{exam_xlsx_csv}
\title{Converts XLSX to CSV}
\usage{
exam_xlsx_csv(XLSX)
}
\arguments{
\item{XLSX}{filename of excel}
}
\value{
filename of CSV
}
\description{
Converts XLSX to CSV
}
\examples{
exam_xlsx_csv("inst/extdata/example-grade-roster.xlsx")
}
