% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_grade_book.R
\docType{data}
\name{example_grade_book}
\alias{example_grade_book}
\title{Example data}
\format{
\subsection{\code{example_grade_book}}{

A data frame with 7 rows and 12 columns:
\describe{
\item{Student}{Student Name}
\item{SIS Login ID}{ID}
\item{OQ1 - Test 2}{Results}
}
}
}
\source{
simulated data
}
\usage{
example_grade_book
}
\description{
An example grade book
}
\keyword{datasets}
