% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_grade_roster.R
\name{fill_grade_roster}
\alias{fill_grade_roster}
\title{fill_grade_roster}
\usage{
fill_grade_roster(IDs, total, infile, outfile, supp = 45:49,
  supp_code = "US10", RP = NULL)
}
\arguments{
\item{IDs}{vector of IDs}

\item{total}{vector of marks - order corresponds to IDs}

\item{infile}{empty grade roster}

\item{outfile}{lace to save complete grade roster}
}
\value{
VOID
}
\description{
Still need to open, check and save
}
\examples{
pacman::p_load(tidyverse, examMarking)
set.seed(2019)
df  <- tibble(
  ID = 1:7,
  total = c(sample(80:100, 4), 0, 49, 44)
)
df
fill_grade_roster(df$ID, df$total,
                  infile = "inst/grade_example.csv",
                  outfile = "inst/example_output.csv")
read_lines("inst/example_output.csv")
fill_grade_roster(df$ID, df$total,
                  infile = "inst/grade_example.csv",
                  outfile = "inst/example_output_2.csv",
                  supp = 40:49)
read_lines("inst/example_output_2.csv")
fill_grade_roster(df$ID, df$total,
                  infile = "inst/grade_example.csv",
                  outfile = "inst/example_output_3.csv",
                  RP = c(2, 5))
read_lines("inst/example_output_3.csv")
}
