% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_email_list.R
\name{get_email_list}
\alias{get_email_list}
\title{Get list of student emails from IDs}
\usage{
get_email_list(IDs)
}
\arguments{
\item{IDs}{vector of IDs of form a1078881 or 1078881}
}
\value{
single string with all emails seperated by a comma
}
\description{
Get list of student emails from IDs
}
\examples{
IDs  <- c('a1078881', 'a1078886')
get_email_list(IDs)
}
