% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grade_summary.R
\name{get_grade_summary}
\alias{get_grade_summary}
\title{Takes column of grades and gives summary}
\usage{
get_grade_summary(grades)
}
\arguments{
\item{grades}{vector of grades}
}
\value{
table of counts and prop.
}
\description{
Takes column of grades and gives summary
}
\examples{
pacman::p_load(examMarking, tidyverse)
data("SMI_2018_marks_clean")
SMI_2018_marks_clean
get_grade_summary(SMI_2018_marks_clean$grade)
}
