% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grades.R
\name{get_grades}
\alias{get_grades}
\title{convert integer marks to grades}
\usage{
get_grades(marks, grade_tab = NULL, trace = FALSE)
}
\arguments{
\item{marks}{vector of integer marks}

\item{grades}{table of grade boundaries, needs cols: grade, lower, upper}
}
\value{
grades
}
\description{
uses a table of upper and lower points to assign
grades
}
\examples{
library(examMarking)
data_frame(marks = 0:100, grade = get_grades(marks)) \%>\% 
 ggplot(aes(marks, grade, col = grade)) + 
 geom_line(size = 4) + 
 scale_x_continuous(breaks = seq(0, 100, 5))
}
