% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean.R
\name{get_mean}
\alias{get_mean}
\title{takes data frame and regular expression and calculates the mean for the given 
columns. Will take into account exemmptions}
\usage{
get_mean(df, RE)
}
\arguments{
\item{df}{data frame of marks}

\item{RE}{regular expression to choose columns to average}
}
\value{
vector of means
}
\description{
takes data frame and regular expression and calculates the mean for the given 
columns. Will take into account exemmptions
}
\examples{
library(examMarking)
data("SMI_2018_marks")
SMI_2018_marks <- 
  SMI_2018_marks \%>\% 
  clean_marks_df("^A") \%>\% 
  get_prop_df("^A", 1)
SMI_2018_marks \%>\% 
  get_mean("^A")
}
