% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop.R
\name{get_prop_df}
\alias{get_prop_df}
\title{Get proportions for dataframe.}
\usage{
get_prop_df(df, assess_RE, total_row)
}
\arguments{
\item{df}{data frame of marks}

\item{assess_RE}{regular expression to choose columns}
}
\value{
data frame with proportions
}
\description{
Identifies columns by regular expression
}
\examples{
library(examMarking)
data("SMI_2018_marks")
SMI_2018_marks  <- 
  SMI_2018_marks \%>\% clean_marks_df("^A")
SMI_2018_marks
SMI_2018_marks \%>\% get_prop_df("^A", 1)
}
