% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total.R
\name{get_total}
\alias{get_total}
\title{takes data frame and regular expression and calculates the totals for the given 
columns}
\usage{
get_total(df, RE)
}
\arguments{
\item{df}{data frame of marks}

\item{RE}{regular expression to choose columns to total}
}
\value{
vector of totals
}
\description{
takes data frame and regular expression and calculates the totals for the given 
columns
}
\examples{
library(examMarking)
data("SMI_2018_marks")
SMI_2018_marks$ET  <- SMI_2018_marks \%>\% get_total("^Q")
}
