% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\docType{data}
\name{marks}
\alias{marks}
\title{Simulated data}
\format{
A data frame with 31 rows and 12 columns
Has a row called total with totals in it.
\describe{
\item{ID}{ID code}
\item{A1}{First assignment etc.}
\item{Q1}{First exam question}
}
}
\source{
sim_data.R
}
\usage{
marks
}
\description{
Simulated data for vignette
}
\keyword{datasets}
