% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{norm_compare}
\alias{norm_compare}
\title{compare norm}
\usage{
norm_compare(raw, norm)
}
\arguments{
\item{raw}{vector of raw marks}

\item{norm}{vector of normalised marks}
}
\value{
list of various comparisons
}
\description{
gives some basic comparions of raw and normed marks
}
\examples{
df  <- data_frame(raw = 0:100, norm = norm_glonek(raw, a = 0.2, b = 2))
norm_compare(df$raw, df$norm)
}
