% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{norm_glonek}
\alias{norm_glonek}
\title{Normalizes using Glonek function}
\usage{
norm_glonek(x, a = 0, b = 1)
}
\arguments{
\item{x}{marks}

\item{a}{normalization parameter}

\item{b}{normalization parameter}
}
\value{
normalized marks
}
\description{
Takes marks and normalizes using Glonek function - linear
scale on logit scale
}
\examples{
data_frame(raw = 0:100, norm = norm_glonek(raw, a = 0.5, b = 2)) \%>\% 
  ggplot(aes(raw, norm)) + geom_point()
}
\author{
Jono Tuke <simon.tuke@adelaide.edu.au>
}
