% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_norm.R
\name{plot_norm}
\alias{plot_norm}
\title{Plot raw and norm}
\usage{
plot_norm(df, raw, norm, ID)
}
\arguments{
\item{df}{data frame of data}

\item{raw}{raw marks}

\item{norm}{normalised marks}

\item{ID}{student ID column}
}
\value{
ggplot
}
\description{
Plot raw and norm
}
\examples{
pacman::p_load(tidyverse)
df  <- data_frame(
  raw = 0:100, 
  norm = norm_glonek(raw, a = 0.2, b = 2), 
  ID = seq_along(raw))
  plot_norm(df, raw, norm, ID)
}
