% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{profile_plot}
\alias{profile_plot}
\title{Plots the exam questions to look for seperation}
\usage{
profile_plot(df, RE, grade)
}
\arguments{
\item{df}{data frame with results}

\item{RE}{regular expression of exam questions}

\item{grade}{column with grades}
}
\value{
a ggplot
}
\description{
Plots the exam questions to look for seperation
}
\examples{
library(examMarking)
data("SMI_2018_marks_clean")
SMI_2018_marks_clean \%>\% profile_plot("^Q", grade)
}
