% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPI.R
\docType{data}
\name{SPI}
\alias{SPI}
\title{Data for previous SPI subject}
\format{
A data frame with 41 rows and 10 columns
\describe{
\item{subject}{Subject code}
\item{year}{Exam year}
\item{sem}{Exam semester}
\item{F}{Fail percent}
\item{P}{Pass percent}
\item{C}{Credit percent}
\item{D}{Distinction percent}
\item{HD}{High Distinction percent}
\item{N}{Number of students}
\item{mean}{Average mark}
}
}
\source{
McCannaliser
}
\usage{
SPI
}
\description{
All the previous SPI results taken from McCannaliser
}
\keyword{datasets}
