% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_marks.R
\name{clean_marks}
\alias{clean_marks}
\title{takes a vector of marks and converts missing to zero and EX to NA}
\usage{
clean_marks(x)
}
\arguments{
\item{x}{marks}
}
\value{
vector of marks
}
\description{
takes a vector of marks and converts missing to zero and EX to NA
}
\examples{
library(examMarking)
data(marks)
clean_marks(marks$A1_20)
}
