% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_calculate_total.R
\name{exam_calculate_total}
\alias{exam_calculate_total}
\title{exam_calculate_total}
\usage{
exam_calculate_total(df, weights)
}
\arguments{
\item{df}{dataframe with proportions for each piece
of asssessment.}

\item{weights}{dataframe with column with assessment names and
column with weights}
}
\value{
vector of total marks.
}
\description{
Uses the proportions for each assessment and the
weightings to get the total for each student. Also
renormalises for exemptions as indicated by NAs.
}
\examples{
pacman::p_load(tidyverse, targets)
df <- tibble(
  A1 = c(1, 0.5, 0),
  A2 = c(0, NA, 1),
  OQ1 = c(1, 1, NA)
)
weights <- tibble(
  assessment = colnames(df),
  weight = c(40, 40, 20)
)
weights
df$total <- exam_calculate_total(df, weights)
df
}
