% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_calculate_total.R
\name{exam_calculate_total}
\alias{exam_calculate_total}
\title{exam_calculate_total}
\usage{
exam_calculate_total(ID, df, weights)
}
\arguments{
\item{ID}{student ID}

\item{df}{marks dataframe}

\item{weights}{tibble with assessment and weight}
}
\value{
tibble with ID and mark
}
\description{
calculates the weighted mark
}
\examples{
pacman::p_load(tidyverse, targets)
df <- tibble(
 ID = LETTERS[1:3],
 A1 = c(1, 0.5, 0),
 A2 = c(0, NA, 1),
 OQ1 = c(1, 1, NA)
)
df
weights_main <- tibble(
 assessment = colnames(df)[-1],
 weight = c(40, 40, 20)
)
weights_C <- tibble(
 assessment = colnames(df)[-1],
 weight = c(50, 50, 0)
)
marks <- LETTERS[1:2] |>
 map_dfr(exam_calculate_total, df, weights_main) |>
 bind_rows(
   exam_calculate_total("C", df, weights)
 )
marks
df |> left_join(marks, "ID")
}
