% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_clean_marks.R
\name{exam_clean_marks}
\alias{exam_clean_marks}
\title{exam_clean_marks}
\usage{
exam_clean_marks(x)
}
\arguments{
\item{x}{column from MyUni gradebook that contains marks}
}
\value{
cleaned column
}
\description{
converts missing to 0, and EX to missing
}
