% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_convert_prop.R
\name{exam_convert_prop}
\alias{exam_convert_prop}
\title{exam_convert_prop}
\usage{
exam_convert_prop(df, cols, row)
}
\arguments{
\item{df}{a dataframe with totals in one row}

\item{cols}{tidyselection of columns}

\item{row}{number of row with totals}
}
\value{
dataframe with proportions
}
\description{
given a dataframe, some columns and a total row, converts
to proportions
}
