% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_fill_grade_roster.r
\name{exam_fill_grade_roster}
\alias{exam_fill_grade_roster}
\title{exam_fill_grade_roster}
\usage{
exam_fill_grade_roster(df, IDs, total, infile, outfile, RAA = 45:49)
}
\arguments{
\item{df}{marks data frame}

\item{IDs}{column that has the studnet IDs}

\item{total}{column that has the final mark}

\item{infile}{empty grade roster}

\item{outfile}{filled grade roster}

\item{RAA}{values to give academic RAA}
}
\value{
filled grade roster
}
\description{
fills the grade roster
}
\examples{
df <- tibble(
  ID = 1:10,
  marks = c(0, 40, 44, 48, 50, 51, 65, 75, 85, 100)
)
exam_fill_grade_roster(
  df, ID, marks,
  "inst/extdata/example-grade-roster.csv",
  "inst/extdata/example-grade-roster-filled.csv"
)
exam_fill_grade_roster(
  df, ID, marks,RAA = 40:49,
  "inst/extdata/example-grade-roster.csv",
  "inst/extdata/example-grade-roster-filled-2.csv"
)
}
