% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_grade_summary.r
\name{exam_grade_summary}
\alias{exam_grade_summary}
\title{Takes column of grades and gives summary}
\usage{
exam_grade_summary(df, grade_col)
}
\arguments{
\item{grades}{vector of grades}
}
\value{
table of counts and prop.
}
\description{
Takes column of grades and gives summary
}
\examples{
pacman::p_load(tidyverse)
grades <- c("FNS", "FNA", "FA", "P", "C", "D", "HD")
df <- tibble(
  grades = sample(grades, 100, replace = TRUE)
)
df |> exam_grade_summary(grades)
}
