% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_plot_grades.r
\name{exam_plot_grades}
\alias{exam_plot_grades}
\title{Get grade plot}
\usage{
exam_plot_grades(df, grade_col)
}
\arguments{
\item{df}{data frame of marks}

\item{grade_col}{column with grades}
}
\value{
ggplot
}
\description{
Get grade plot
}
\examples{
# pacman::p_load(tidyverse)
grades <- c("FNS", "FNA", "FA", "P", "C", "D", "HD")
df <- tibble(
  grades = sample(grades, 100, replace = TRUE)
)
df |> exam_grade_summary(grades)
df |> exam_plot_grades(grades)
}
