% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop.R
\name{get_prop_vec}
\alias{get_prop_vec}
\title{Calculates the proportion for each mark in a vector.}
\usage{
get_prop_vec(x, total_row)
}
\arguments{
\item{x}{column of marks with total in name of vector}
}
\value{
vector of proportions.
}
\description{
Assumes that colname has total after _
}
\examples{
library(examMarking)
data("SMI_2018_marks")
SMI_2018_marks  <- SMI_2018_marks \%>\% clean_marks_df("^A")
get_prop_vec(SMI_2018_marks$`Assignment 1 (77417)`, 1)
}
